test: scaling in the model, not in the C code

+++ App/img_array.h
@@ -1,15 +1,17 @@
 #ifndef img_arrayH
 #define img_arrayH
 
-extern const float img_array0[];
-extern const float img_array1[];
-extern const float img_array2[];
-extern const float img_array3[];
-extern const float img_array4[];
-extern const float img_array5[];
-extern const float img_array6[];
-extern const float img_array7[];
-extern const float img_array8[];
-extern const float img_array9[];
+#include <stdint.h>
+
+extern const uint8_t img_array0[];
+extern const uint8_t img_array1[];
+extern const uint8_t img_array2[];
+extern const uint8_t img_array3[];
+extern const uint8_t img_array4[];
+extern const uint8_t img_array5[];
+extern const uint8_t img_array6[];
+extern const uint8_t img_array7[];
+extern const uint8_t img_array8[];
+extern const uint8_t img_array9[];
 
 #endif
